Another round of TSTP integration. Some minor bug fixes: unsigned missing at Hash, and some reinterpret_cast replaced by static_cast. Incorporation of Mateus fixed E100 (yet with too many noise on the output).

+++ include/architecture/armv7/pmu.h
@@ -111,7 +111,7 @@ public:
     ARMv7_A_PMU() {}
 
     static void config(const Channel & channel, const Event & event, const Flags & flags = NONE) {
-        assert((channel < CHANNELS) && (event < EVENTS));
+        assert((static_cast<unsigned int>(channel) < CHANNELS) && (static_cast<unsigned int>(event) < EVENTS));
         db<PMU>(TRC) << "PMU::config(c=" << channel << ",e=" << event << ",f=" << flags << ")" << endl;
         pmselr(channel);
         pmxevtyper(event);
